Feature/108112

+++ SME.CDEP.Dominio/Entidades/AcervoDocumentalCompleto.cs
@@ -17,8 +17,8 @@ public class AcervoDocumentalCompleto : EntidadeBaseAuditavel
     public string NumeroPagina { get; set; }
     public string Volume { get; set; }
     public string TipoAnexo { get; set; }
-    public float? Largura { get; set; }
-    public float? Altura { get; set; }
+    public double? Largura { get; set; }
+    public double? Altura { get; set; }
     public string TamanhoArquivo { get; set; }
     public string Localizacao { get; set; }
     public bool CopiaDigital { get; set; }
